requirements.txt
azure-ai-voicelive
azure-identity
azure-core
python-dotenv
pyaudio
fastapi
uvicorn
pydantic
pydantic-settings
supabase